# openapi.yaml
openapi: 3.1.0
info:
  title: Specht API
  version: 1.0.0
  summary: Vulnerability management for SCA, SAST, and IaC scan results.
  description: |
    Specht ingests scanner reports, tracks findings across scans, evaluates a
    deployment gate, and manages waivers, policy, teams, and threat intel.

    Everything lives under `/api/v1`. Authentication is always an
    `Authorization: Bearer <credential>` header carrying either a session JWT
    (from `/auth/login`, `/auth/refresh`, or the SSO callback) or a
    project-scoped API key (prefix `vuln_`, from `/auth/apikeys`).

    ## Authorization model

    Three gates apply, in order:

    1. **Credential type** — some routes require a session and reject API keys
       entirely (`x-auth: session`), because a project key cannot represent the
       authority they carry.
    2. **Role or scope** — session roles are `admin`, `editor`, `viewer`
       (`x-required-role`); API key scopes are `ingest`, `read`, `admin`
       (`x-required-scope`). Sessions pass scope checks and are then subject to
       project membership instead.
    3. **Project access** — the caller must be a member of the project (session)
       or the key must be bound to it (API key), otherwise the request fails
       with `403 project_access_denied`, or `404 not_found` when an admin asks
       about a project that does not exist.

    ## Conventions

    - List endpoints return a **bare JSON array**; pagination is
...
      total is unknown, not zero.
...
servers:
...
security: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Create an account
      description: |
        Public signup. The first account created is promoted to admin by the
        `ADMIN_EMAILS` bootstrap on the next boot, not by this endpoint.
        A duplicate email is indistinguishable from any other failure, by
        design, so this cannot enumerate accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailPasswordRequest" }
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "413": { $ref: "#/components/responses/BodyTooLarge" }
        "422":
          description: Registration failed (generic).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: { code: registration_failed, message: registration failed } }
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Exchange credentials for tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailPasswordRequest" }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            Invalid email or password. Unknown accounts are deliberately
            indistinguishable from wrong passwords.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: { code: login_failed, message: invalid email or password } }
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Rotate a refresh token
      description: |
        Returns a new access token and a new refresh token. The presented
        refresh token is invalidated; presenting it again revokes the whole
        token family, so a stolen token cannot be replayed silently.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: Rotated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid, expired, revoked, or replayed refresh token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: { code: refresh_failed, message: invalid refresh token } }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Revoke a session
      description: |
        Revokes the refresh token family. When a bearer access token is also
        supplied, that token is retired immediately instead of waiting for its
        normal expiry. Works with only a refresh token in the body.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/auth/sso/callback:
    get:
      tags: [Authentication]
      operationId: ssoCallback
      summary: Complete the SSO flow
      description: |
        Exchanges the authorization code, validates the ID token (signature,
        issuer, audience, expiry, nonce) against the state cookie, resolves the
        IdP subject to a local account, and redirects to the SPA with the
        session token in the URL **fragment** so it never reaches a server log
        or `Referer` header.

        Unknown IdP subjects are provisioned only when their email domain is in
        `SSO_ALLOWED_DOMAINS`; otherwise the request is refused. An existing
        account keeps its local role regardless of IdP group membership.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "302":
          description: "Redirect to `/#sso_token=<jwt>`."
          headers:
            Location:
              schema: { type: string }
        "400":
          description: Missing code, or the state cookie did not match.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: |
            The IdP subject is not provisioned and its email domain is not
            allowlisted, or the asserted email is empty.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Token exchange or identity extraction failed.
  /api/v1/me:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: Read the signed-in account
      x-auth: session
      security:
        - sessionAuth: []
      responses:
        "200":
          description: The current profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SessionRequired" }
    put:
      tags: [Profile]
      operationId: updateProfile
      summary: Update the signed-in account
      description: |
        Updates the display name. The field is a pointer: omit it to leave the
        stored name alone, send `""` to clear it.
      x-auth: session
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdateRequest" }
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SessionRequired" }
  /api/v1/auth/apikeys:
    post:
      tags: [API keys]
      operationId: createApiKey
      summary: Mint a project API key
      description: |
        Creates a key bound to exactly one project. The secret is returned once,
        in `raw_key`, and never again — later reads expose only the prefix and
        the last four characters.

        Reachable by an admin session, or by an API key that itself holds the
        `admin` scope.
      x-required-role: admin
      x-required-scope: admin
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateApiKeyRequest" }
      responses:
        "201":
          description: Key created. `raw_key` appears only here.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKey" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    get:
      tags: [API keys]
      operationId: listApiKeys
      summary: List API keys
      description: Lists the caller's key set with the secret masked.
      x-required-role: admin
      x-required-scope: admin
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: The keys.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/apikeys/{id}:
    delete:
      tags: [API keys]
      operationId: revokeApiKey
      summary: Revoke an API key
      x-required-role: admin
      x-required-scope: admin
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /api/v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List visible projects
      description: |
        Scoped to the caller's identity: an admin sees every project, an API key
        sees exactly the project it is bound to, and a member sees only the
        projects they belong to.
      x-required-scope: read
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: The visible projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      description: Session admins only; a project key can never create tenancy.
      x-auth: session
      x-required-role: admin
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SessionRequired" }
        "409":
          description: The slug is taken.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: { code: slug_taken, message: project slug already exists } }
  /api/v1/projects/{slug}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Read a project
      x-required-scope: read
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ProjectAccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Projects]
      operationId: updateProject
      summary: Update a project
      description: Both fields are pointers; omit one to leave it unchanged.
      x-required-scope: admin
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectRequest" }
      responses:
        "200":
          description: The updated project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete a project
      description: Deletes the project and everything hanging off it. Returns the deleted project.
      x-required-scope: admin
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: The project as it was before deletion.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403": { $ref: "#/components/responses/ProjectAccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/scanners:
    get:
      tags: [Scanners]
      operationId: listScanners
      summary: List the scanner catalog
      description: |
        Every registered scanner with its version, the finding kinds it emits,
        the scan types it accepts, and whether it can detect its own format.
        Session admins only — project keys are refused.
      x-auth: session
      x-required-role: admin
      security:
        - sessionAuth: []
      responses:
        "200":
          description: The catalog, in registration order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Scanner" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SessionRequired" }
  /api/v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List accounts
      description: |
        The directory an admin picks from when granting project membership or
        adding a team member, so a UI never has to ask for a raw user id.
        Session admins only — a project-scoped key has no business
        enumerating the organization. The response carries no credential
        material.
      x-auth: session
      x-required-role: admin
      security:
        - sessionAuth: []
      parameters:
        - name: email
          in: query
          description: Case-insensitive substring match on the email address. Empty lists every account.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Accounts, ordered by email so pages are stable.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SessionRequired" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/projects/{slug}/reports:
    get:
      tags: [Reports]
      operationId: listReports
      summary: List a project's reports
      x-required-scope: read
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Slug"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: The reports, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/reports/{id}:
    get:
      tags: [Reports]
      operationId: getReport
      summary: Read a report
      description: A foreign report is indistinguishable from a missing one.
      x-required-scope: read
      security:
        - sessionAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session JWT from `/auth/login`, `/auth/refresh`, or the SSO callback.
        Carries a role: `admin`, `editor`, or `viewer`. Rejected on routes
        marked `x-auth: session` only when the credential is actually an API
        key — a session is always accepted there.
    apiKeyAuth:
      type: http
      scheme: bearer
      description: |
        Project-scoped API key from `/auth/apikeys`, always prefixed `vuln_`.
        Carries any of the scopes `ingest`, `read`, `admin`, and is bound to one
        project, so it can never read another. Routes marked `x-auth: session`
        refuse it outright.
...